+++ src/grepros/api.py
@@ -8,7 +8,7 @@ Released under the BSD License.
 
 @author      Erki Suurjaak
 @created     01.11.2021
-@modified    29.12.2023
+@modified    30.01.2024
 ------------------------------------------------------------------------------
 """
 ## @namespace grepros.api
@@ -1001,7 +1001,7 @@ def dict_to_message(dct, msg):
         v, msgv = dct[name], realapi.get_message_value(msg, name, typename)
 
         if realapi.is_ros_message(msgv):
-            v = v if is_ros_time(v) and is_ros_time(msgv) else dict_to_message(v, msgv)
+            v = v if is_ros_message(v) else dict_to_message(v, msgv)
         elif isinstance(msgv, (list, tuple)):
             scalarname = realapi.scalar(typename)
             if scalarname in ROS_BUILTIN_TYPES:
@@ -1009,7 +1009,7 @@ def dict_to_message(dct, msg):
                 v = [x if isinstance(x, cls) else cls(x) for x in v]
             else:
                 cls = realapi.get_message_class(scalarname)
-                v = [dict_to_message(x, cls()) for x in v]
+                v = [x if realapi.is_ros_message(x) else dict_to_message(x, cls()) for x in v]
         else:
             v = type(msgv)(v)